Bugfix for a bug I just created; needed to be uint8 not int16.
git-svn-id: https://svn.code.sf.net/p/kageki/code/libmsocket@22 10193a71-e229-423b-a074-d0e0c923123e

+++ socket.c
@@ -104,7 +104,7 @@ inline MSocket *lms_socket_findbyfd(int fd)
  * type = the type of the new socket to allocate
  *
  */
-MSocket *lms_socket_create(int16_t type)
+MSocket *lms_socket_create(uint8_t type)
 {
 	MSocket *ptr;
